* develop: (303 commits)
  changed Dockerfile and Jenkins for production
  fixing rc-slider
  fixing map
  fixing tooltip issue
  Datasets: changelog
  Datasets: geoinfo always added. if raster always true
  update Change log file
  Changelog updated
  use interactionConfig for Carto layers tooltips
  fix tooltip for Carto layers
  update how CartoDB layers are fetched, we use pulseConfig now
  added command to check layers integrity
  Vertically-responsive bar charts (#298)
  make getmarkerheight static
  clear topics filter selection when clicking on topics tag
  a few fixes: 1 planet pulse layers definition, explore filters result, datasets search page size
  fix in TreeUtil
  show labels instead of ids in the selectors sum up text. Move tags icon
  up changelog
  added support for gee and moving dependencies to use widget editor components
  ...

+++ components/datasets/list/DatasetsListCard.js
@@ -22,7 +22,8 @@ class DatasetsListCard extends React.Component {
   @Autobind
   handleDelete() {
     const { dataset } = this.props;
-    toastr.confirm(`Are you sure you want to delete the dataset: ${dataset.name}?`, {
+    const metadata = dataset.metadata[0];
+    toastr.confirm(`Are you sure you want to delete the dataset: ${metadata && metadata.attributes.info ? metadata.attributes.info.name : dataset.name}?`, {
       onOk: () => {
         this.service.deleteData(dataset.id)
           .then(() => {
@@ -36,7 +37,7 @@ class DatasetsListCard extends React.Component {
 
   render() {
     const { dataset, routes } = this.props;
-
+    const metadata = dataset.metadata[0];
     return (
       <div className="c-card c-datasets-list-card">
         <div className="card-container">
@@ -47,7 +48,7 @@ class DatasetsListCard extends React.Component {
             >
               <a>
                 <Title className="-default">
-                  {dataset.name}
+                  {metadata && metadata.attributes.info ? metadata.attributes.info.name : dataset.name}
                 </Title>
               </a>
             </Link>
